update libmpeg2 update libtheora update x264/Makefile update xvidcore/Makefile
git-svn-id: https://ffdshow-tryout.svn.sourceforge.net/svnroot/ffdshow-tryout/trunk@1985 3b938f2f-1a1a-0410-8054-a526ea5ff92c

+++ src/xiph/ogg/ogg.h
@@ -25,7 +25,7 @@ extern "C" {
 
 typedef struct {
   void *iov_base; 
-  size_t iov_len;
+  unsigned iov_len;
 } ogg_iovec_t;
 
 typedef struct {
